Extract shared dock-trigger logic into a hook
StatefulDockStart/StatefulDockEnd duplicated the same useDockReservation, isDisabled, and dockAction-dispatch wiring; moved into a new useDockTrigger(slot, associatedKey) hook
